# Qt5: improve repeated zooming performance